+++ packages/auth/amplify_auth_cognito/example/integration_test/mfa_totp_optional_test.dart
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import 'package:amplify_auth_cognito_dart/amplify_auth_cognito_dart.dart';
 import 'package:amplify_auth_integration_test/amplify_auth_integration_test.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:amplify_integration_test/amplify_integration_test.dart';
@@ -93,6 +94,65 @@ void main() {
           'w/ device name',
           (_) => runTest(friendlyDeviceName: friendlyDeviceName),
         );
+
+        asyncTest('verifyTotpSetup allows retries', (_) async {
+          final username = generateUsername();
+          final password = generatePassword();
+
+          await adminCreateUser(
+            username,
+            password,
+            verifyAttributes: true,
+            autoConfirm: true,
+          );
+
+          final signInRes = await Amplify.Auth.signIn(
+            username: username,
+            password: password,
+          );
+
+          check(
+            because: 'TOTP is optional',
+            signInRes.nextStep.signInStep,
+          ).equals(AuthSignInStep.done);
+
+          final totpSetupResult = await Amplify.Auth.setUpTotp();
+
+          try {
+            await Amplify.Auth.verifyTotpSetup('555555');
+            fail('Expected to fail');
+          } on AuthException catch (e) {
+            check(
+              e,
+              because: 'Invalid TOTP code should fail verification',
+            ).isA<EnableSoftwareTokenMfaException>();
+          }
+
+          check(
+            await cognitoPlugin.fetchMfaPreference(),
+            because: 'TOTP should not be enabled',
+          ).equals(
+            const UserMfaPreference(
+              enabled: {},
+              preferred: null,
+            ),
+          );
+
+          try {
+            await Amplify.Auth.verifyTotpSetup(
+              await generateTotpCode(totpSetupResult.sharedSecret),
+            );
+          } on Exception catch (e) {
+            fail('Expected to succeed, but got $e');
+          }
+
+          check(await cognitoPlugin.fetchMfaPreference()).equals(
+            const UserMfaPreference(
+              enabled: {MfaType.totp},
+              preferred: MfaType.totp,
+            ),
+          );
+        });
       });
     });
   });