+++ packages/auth/amplify_auth_cognito_test/lib/common/mock_legacy_credential_provider.dart
@@ -30,4 +30,20 @@ class MockLegacyCredentialProvider implements LegacyCredentialProvider {
   }) async {
     return data;
   }
+
+  @override
+  Future<LegacyDeviceDetails?> fetchLegacyDeviceSecrets({
+    required String username,
+    CognitoUserPoolConfig? userPoolConfig,
+  }) async {
+    return null;
+  }
+
+  @override
+  Future<void> deleteLegacyDeviceSecrets({
+    required String username,
+    CognitoUserPoolConfig? userPoolConfig,
+  }) async {
+    return;
+  }
 }